refactor node data
